ROFO-104 FoodSearchService 역할 위임
* FoodQueryService에서 가게 조회
* UserCommandService에서 코인 감소
